data_pipeline/requirements.txt
pandas==2.1.0
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9